Gather wallet actions into a properly-named file
